+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/feat/bot/DiscordBot.kt
@@ -6,6 +6,7 @@ import dev.kord.common.entity.Snowflake
 import dev.kord.core.Kord
 import dev.kord.core.event.gateway.DisconnectEvent
 import dev.kord.core.event.gateway.ResumedEvent
+import dev.kord.core.event.interaction.AutoCompleteInteractionCreateEvent
 import dev.kord.core.event.interaction.ButtonInteractionCreateEvent
 import dev.kord.core.event.interaction.GuildChatInputCommandInteractionCreateEvent
 import dev.kord.core.event.message.MessageCreateEvent
@@ -16,15 +17,15 @@ import io.github.aakira.napier.Napier
 import io.github.sophon.core.architecture.onError
 import io.github.sophon.core.featureConfig.model.Config
 import io.github.sophon.discord.feat.admin.adminCommands
+import io.github.sophon.discord.feat.bot.usecase.HandleAutoCompleteEventUseCase
 import io.github.sophon.discord.feat.bot.usecase.HandleButtonInteractionUseCase
+import io.github.sophon.discord.feat.bot.usecase.HandleQueryUseCase
 import io.github.sophon.discord.feat.bot.usecase.PostDailyReportEmbedUseCase
-import io.github.sophon.discord.feat.bot.usecase.ResultToEmbedUseCase
-import io.github.sophon.discord.feat.bot.usecase.RouteCommandToFeatureUseCase
 import io.github.sophon.discord.feat.config.BotFeatureRepo
 import io.github.sophon.discord.feat.core.domain.Tracker
 import io.github.sophon.discord.feat.core.domain.model.BotOutput
+import io.github.sophon.discord.feat.core.domain.model.Command.Argument.AutoCompleteType
 import io.github.sophon.discord.util.safeRestCall
-import io.github.sophon.integration.model.Source
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.collectLatest
@@ -40,8 +41,8 @@ internal class DiscordBotImpl(
     private val kord: Kord,
     private val tracker: Tracker,
     private val adminConfig: Config.AdminConfig,
-    private val routeCommandToFeatureUseCase: RouteCommandToFeatureUseCase,
-    private val resultToEmbedUseCase: ResultToEmbedUseCase,
+    private val handleQueryUseCase: HandleQueryUseCase,
+    private val handleAutoCompleteEventUseCase: HandleAutoCompleteEventUseCase,
     private val handleButtonInteractionUseCase: HandleButtonInteractionUseCase,
     private val postDailyReportEmbedUseCase: PostDailyReportEmbedUseCase,
     private val coroutineScope: CoroutineScope,
@@ -73,22 +74,20 @@ internal class DiscordBotImpl(
         monitorGatewayHealth()
 
         kord.on<GuildChatInputCommandInteractionCreateEvent> {
-            safeRestCall(TAG) { handleCommand() }
+            handleQueryUseCase.invoke(
+                interaction = interaction,
+                editableEmbedMap = editableEmbedMap,
+            )
         }
 
         kord.on<MessageCreateEvent> {
-            // ignoring other bots, even ourselves
-            if (message.author?.isBot != false) return@on
-
-            // ignoring if someone replies with tag
-            val botId = kord.selfId
-            val botMention = "<@$botId>"
-            val botNicknameMention = "<@!$botId>"
-            if (botMention !in message.content && botNicknameMention !in message.content) {
-                return@on
+            safeRestCall(TAG) {
+                handleQueryUseCase.invoke(
+                    message = message,
+                    botId = kord.selfId,
+                    editableEmbedMap = editableEmbedMap,
+                )
             }
-
-            safeRestCall(TAG) { handleMessage() }
         }
 
         kord.on<ButtonInteractionCreateEvent> {
@@ -96,6 +95,12 @@ internal class DiscordBotImpl(
                 .onError { error ->
                     Napier.e(tag = TAG) { "${interaction.data.guildId} → Button interaction: $error" }
                 }
+        }
+
+        kord.on<AutoCompleteInteractionCreateEvent> {
+            safeRestCall(TAG) {
+                handleAutoCompleteEventUseCase.invoke(interaction)
+            }
         }
 
         //‼️ THIS SUSPENDS UNTIL LOGGED OUT
@@ -129,49 +134,6 @@ internal class DiscordBotImpl(
         Napier.e(tag = TAG, throwable = e) { "Failed to delete old commands" }
     }
 
-    private suspend fun MessageCreateEvent.handleMessage() {
-        if (kord.selfId !in message.mentionedUserIds) return
-
-        val source = Source(
-            username = message.author?.username.orEmpty(),
-            id = message.author?.id.toString(),
-            channelId = message.channelId.toString(),
-            serverName = message.getGuildOrNull()?.name.orEmpty(),
-        )
-
-        val result = routeCommandToFeatureUseCase.invoke(
-            source = source,
-            message = message.content,
-        )
-
-        with (resultToEmbedUseCase) {
-            invoke(source, result, coroutineScope, editableEmbedMap)
-        }
-    }
-
-    private suspend fun GuildChatInputCommandInteractionCreateEvent.handleCommand() {
-        val commandString = interaction.command.rootName
-            .lowercase()
-        val query = interaction.command.strings.values
-            .joinToString(" ")
-        val source = Source(
-            username = interaction.user.username,
-            id = interaction.user.data.id.toString(),
-            channelId = interaction.channelId.toString(),
-            serverName = interaction.getGuildOrNull()?.name.orEmpty(),
-        )
-
-        val result = routeCommandToFeatureUseCase.invoke(
-            source = source,
-            commandString = commandString,
-            query = query
-        )
-
-        with (resultToEmbedUseCase) {
-            invoke(source, result, coroutineScope, editableEmbedMap)
-        }
-    }
-
     @Suppress("UnusedPrivateMember")
     private suspend fun createCommandsForTestServer() {
         val testGuildSnowFlake = Snowflake(adminConfig.adminServerId)
@@ -192,6 +154,7 @@ internal class DiscordBotImpl(
                         supportedCommand.argumentList.forEach { argument ->
                             string(name = argument.name, description = argument.description) {
                                 required = argument.isRequired
+                                autocomplete = (argument.autoCompleteType != AutoCompleteType.None)
                             }
                         }
                     }
@@ -217,6 +180,7 @@ internal class DiscordBotImpl(
                             supportedCommand.argumentList.forEach { argument ->
                                 string(name = argument.name, description = argument.description) {
                                     required = argument.isRequired
+                                    autocomplete = (argument.autoCompleteType != AutoCompleteType.None)
                                 }
                             }
                         }
@@ -241,6 +205,7 @@ internal class DiscordBotImpl(
                         command.argumentList.forEach { argument ->
                             string(name = argument.name, description = argument.description) {
                                 required = argument.isRequired
+                                autocomplete = (argument.autoCompleteType != AutoCompleteType.None)
                             }
                         }
                     }