+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/feat/featureRegistryModule.kt
@@ -24,7 +24,9 @@ import io.github.sophon.discord.feat.bot.usecase.CreateMutableEmbedUseCase
 import io.github.sophon.discord.feat.bot.usecase.CreatePlainMessageUseCase
 import io.github.sophon.discord.feat.bot.usecase.CreateReplyEmbedUseCase
 import io.github.sophon.discord.feat.bot.usecase.GetBotFeatureInfoUseCase
+import io.github.sophon.discord.feat.bot.usecase.HandleAutoCompleteEventUseCase
 import io.github.sophon.discord.feat.bot.usecase.HandleButtonInteractionUseCase
+import io.github.sophon.discord.feat.bot.usecase.HandleQueryUseCase
 import io.github.sophon.discord.feat.bot.usecase.PostDailyReportEmbedUseCase
 import io.github.sophon.discord.feat.bot.usecase.ResultToEmbedUseCase
 import io.github.sophon.discord.feat.bot.usecase.RouteCommandToFeatureUseCase
@@ -114,6 +116,8 @@ internal val featureRegistryModule = module {
     singleOf(::CreateReplyEmbedUseCase)
     singleOf(::ResultToEmbedUseCase)
     singleOf(::CreateMutableEmbedUseCase)
+    singleOf(::HandleQueryUseCase)
+    singleOf(::HandleAutoCompleteEventUseCase)
     singleOf(::HandleButtonInteractionUseCase)
     singleOf(::PostDailyReportEmbedUseCase)
     //endregion