+++ bot/discord/src/jvmMain/kotlin/io/github/sophon/discord/feat/wikiWavu/usecase/SearchStringFollowupsUseCase.kt
@@ -29,12 +29,12 @@ internal class SearchStringFollowupsUseCase(
         val startingMoveInput = parts.drop(1).joinToString()
         return getMovesUseCase.invoke(
             wiki = wiki,
-            charName = charName,
+            characterId = charName,
             filter = TekkenFilters.Strings(startingMoveInput),
-        ).map { moveList ->
+        ).map { (character, moveList) ->
             BotOutput(
                 primaryEmbedBuilder = moveListEmbed(
-                    category = "${query.uppercase()} Followups",
+                    category = "${character.displayName.uppercase()} ${startingMoveInput.uppercase()} Followups",
                     dataList = moveList.map {
                         "${it.input}: ${it.guard}"
                     },
@@ -42,7 +42,7 @@ internal class SearchStringFollowupsUseCase(
                     color = Color(BLUE),
                 ),
                 buttons = BotOutput.ButtonSet(
-                    buttonList = moveList.toButtons(charName),
+                    buttonList = moveList.toButtons(charName = character.id),
                     duration = EMBED_BUTTON_DURATION_INF.seconds,
                 ),
             )